Move sender thread object to own file
